Moved search into its own class
